Fix the name and GitHub link of the project
Born as a project to read/write the bootcount on the TI am3xx platform,
over time it has introduced bootcount management on EEPROM and for the
stm32mp1 platform. As a result, the project removed the 'davinci' tag
from its name and GitHub link. The patch fixes references to the old
project name that are still present in some project files.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>